pyproject.toml
[build-system]
requires = ["setuptools>=67"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp_country_server"
version = "0.1.0"
description = "MCP-style tool server scaffold"
authors = [{ name = "Demo Author" }]
requires-python = ">=3.10"
dependencies = [
  "Flask>=3.0.3",
  "requests>=2.32.3",
  "jsonschema>=4.23.0",
  "python-dotenv>=1.0.1"
]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"
